Implementation du responsive actualite: fix responsive
