+++ front/src/components/business/FilterGroup/FilterBarMobile.tsx
@@ -3,10 +3,9 @@ import React from 'react';
 import cs from 'classnames';
 import { useForm } from 'react-hook-form';
 
-import { FilterHistoryItemProps } from '../FilterHistory';
 import { FilterHistory } from '../FilterHistory/FilterHistory';
-import { SearchDisplayType } from '../Search';
 import { FilterGroup } from './FilterGroup';
+import { FilterBarForm, FilterBarProps } from './types';
 import { Button } from '@/components/common/Button';
 import { Icon } from '@/components/common/Icon';
 import { SvgIcon } from '@/components/common/SvgIcon';
@@ -14,34 +13,7 @@ import { Typography } from '@/components/common/Typography';
 import { PeriodInput } from '@/components/input/StandardInputs/PeriodInput';
 import { FILTER_MEANS_OF_TRANSPORT, FILTER_NEWS_TAGS, FILTER_STATUS } from '@/utils/filters';
 
-export type FilterBarProps = {
-  title: string;
-  searchType?: SearchDisplayType;
-  noSearchBar?: boolean;
-  noPeriodInput?: boolean;
-  filterHistories?: FilterHistoryItemProps[];
-  filtersCount?: number;
-  onValidateFilter: (data: FilterBarForm) => void;
-  withMeanOfTransportFilter?: boolean;
-  withStatusFilter?: boolean;
-  withNewsTagsFilter?: boolean;
-  defaultSearchValue?: string;
-  isMobile?: boolean;
-  filterBarData?: FilterBarForm;
-  open: boolean;
-  setOpen: (open: boolean) => void;
-};
-
-export interface FilterBarForm {
-  status?: string[];
-  meanOfTransport?: string[];
-  newsTags?: string[];
-  startDate?: Date | null;
-  endDate?: Date | null;
-  search: string | null;
-}
-
-export const FilterBar = ({
+export const FilterBarMobile = ({
   title = 'Plus de filtres',
   noSearchBar = false,
   noPeriodInput = false,
@@ -51,7 +23,6 @@ export const FilterBar = ({
   withStatusFilter = false,
   withNewsTagsFilter = false,
   defaultSearchValue,
-  isMobile = true,
   onValidateFilter,
   filterBarData,
   open,
@@ -83,7 +54,7 @@ export const FilterBar = ({
     setEndFocused(isFocused);
   };
 
-  return isMobile ? (
+  return (
     <div className="flex flex-col rounded-xl bg-gray-100 px-5 py-4">
       <div
         className="grid cursor-pointer grid-cols-[20px_1fr_20px] items-center justify-items-center gap-2"
@@ -144,7 +115,6 @@ export const FilterBar = ({
                 control={control}
                 name="meanOfTransport"
                 filters={FILTER_MEANS_OF_TRANSPORT}
-                isMobile={isMobile}
               />
             )}
             {withStatusFilter && (
@@ -153,7 +123,6 @@ export const FilterBar = ({
                 control={control}
                 name="status"
                 filters={FILTER_STATUS}
-                isMobile={isMobile}
               />
             )}
             {withNewsTagsFilter && (
@@ -162,7 +131,6 @@ export const FilterBar = ({
                 control={control}
                 name="newsTags"
                 filters={FILTER_NEWS_TAGS}
-                isMobile={isMobile}
               />
             )}
           </div>
@@ -174,75 +142,6 @@ export const FilterBar = ({
           </div>
         </form>
       </div>
-    </div>
-  ) : (
-    <div className="flex flex-col rounded-xl bg-gray-100 p-5 max-w-[781px]">
-      <form onSubmit={handleSubmit(onSubmit)}>
-        <div className="flex flex-row items-start gap-5">
-          <div className="flex flex-col gap-4">
-            <div className="flex flex-row justify-between items-center gap-4">
-              {!noSearchBar && (
-                <input
-                  data-testid="input-search-element"
-                  enterKeyHint="search"
-                  placeholder="Thèmes, mots-clés..."
-                  className="block w-full rounded-full py-2 px-5 text-base placeholder:font-light placeholder:text-xs placeholder:italic placeholder:text-secondary-400 focus:border-secondary-300 focus:outline-none focus:ring-transparent border-none"
-                  {...register('search')}
-                />
-              )}
-              {!noPeriodInput && (
-                <PeriodInput
-                  noBorder
-                  endDateName="endDate"
-                  startDateName="startDate"
-                  control={control}
-                  isStartFocused={onFocusedStartPeriodInput}
-                  isEndFocused={onFocusedEndPeriodInput}
-                />
-              )}
-            </div>
-            {(startFocused || endFocused) && <div className="h-[260px]" />}
-            {withMeanOfTransportFilter && (
-              <FilterGroup
-                title="Moyen de transport"
-                control={control}
-                name="meanOfTransport"
-                filters={FILTER_MEANS_OF_TRANSPORT}
-                isMobile={isMobile}
-              />
-            )}
-            {withStatusFilter && (
-              <FilterGroup
-                title="Statut de la déclaration"
-                control={control}
-                name="status"
-                filters={FILTER_STATUS}
-                isMobile={isMobile}
-              />
-            )}
-            {withNewsTagsFilter && (
-              <FilterGroup
-                title="Filter par"
-                control={control}
-                name="newsTags"
-                filters={FILTER_NEWS_TAGS}
-                isMobile={isMobile}
-              />
-            )}
-          </div>
-          <div className="flex flex-col gap-8">
-            <button
-              type="submit"
-              className="flex flex-row gap-5 bg-primary-600 rounded-full text-white items-center justify-center px-5 py-2.5"
-            >
-              <Typography size="text-2xs" color="white">
-                Rechercher
-              </Typography>
-              <Icon name="search" size="sm" color="white" />
-            </button>
-          </div>
-        </div>
-      </form>
     </div>
   );
 };
