+++ front/tailwind.config.js
@@ -9,9 +9,6 @@ module.exports = {
   content: [path.join(__dirname, './src/**/*.(js|jsx|ts|tsx)')],
   safelist: ['truncate'],
   theme: {
-    screens: {
-      laptop: '1121px',
-    },
     extend: {
       fontSize: {
         '2xs': '.625rem',
